# docs/POCKETOS_CLIENT_API.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pocket OS Client API",
    "version": "1",
    "description": "Canonical versioned client contract consumed by the web console and the future iPhone application."
  },
  "paths": {
...
    "/api/v1/replay/inspect": {
      "get": {
        "summary": "V1 Replay Inspect",
        "operationId": "v1_replay_inspect_api_v1_replay_inspect_get",
        "parameters": [
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "include_decisions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Decisions"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response V1 Replay Inspect Api V1 Replay Inspect Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/decisions/propose": {
      "post": {
        "summary": "V1 Decisions Propose",
        "operationId": "v1_decisions_propose_api_v1_decisions_propose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response V1 Decisions Propose Api V1 Decisions Propose Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProposalBody": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "decision_id": {
            "type": "string",
            "title": "Decision Id",
            "default": ""
          },
          "risk": {
            "type": "string",
            "title": "Risk",
            "default": "MEDIUM"
          },
          "reversible": {
            "type": "boolean",
            "title": "Reversible",
            "default": true
          },
          "send_to_council": {
            "type": "boolean",
            "title": "Send To Council",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ProposalBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}